Add Feathercoin Icon
For the portable and standalone build Windows uses the icon
set by Qt. Therefore, let PyQt5 use a light version for Windows
in order to increase the contrast with the dark Windows taskbar.
Also, since Gnome 3 (default on Ubuntu) also has a dark theme
we will use the light icon there, too. To be consistent also
make it light on macOS.